+++ src/main/java/com/robertx22/age_of_exile/aoe_data/database/player_skills/PlayerSkillsAdder.java
@@ -3,6 +3,7 @@ package com.robertx22.age_of_exile.aoe_data.database.player_skills;
 import com.robertx22.age_of_exile.aoe_data.database.player_skills.skills.crafting.AlchemyAdder;
 import com.robertx22.age_of_exile.aoe_data.database.player_skills.skills.crafting.CookingAdder;
 import com.robertx22.age_of_exile.aoe_data.database.player_skills.skills.crafting.InscribingAdder;
+import com.robertx22.age_of_exile.aoe_data.database.player_skills.skills.crafting.TinkeringAdder;
 import com.robertx22.age_of_exile.aoe_data.database.player_skills.skills.gathering.ExplorationAdder;
 import com.robertx22.age_of_exile.aoe_data.database.player_skills.skills.gathering.FarmingAdder;
 import com.robertx22.age_of_exile.aoe_data.database.player_skills.skills.gathering.FishingAdder;
@@ -14,6 +15,7 @@ public class PlayerSkillsAdder implements ISlashRegistryInit {
     @Override
     public void registerAll() {
 
+        TinkeringAdder.createSkill();
         ExplorationAdder.createSkill();
         MiningAdder.createSkill();
         FarmingAdder.createSkill();