+++ src/main/java/com/robertx22/age_of_exile/database/base/CreativeTabs.java
@@ -11,6 +11,7 @@ import com.robertx22.age_of_exile.vanilla_mc.items.gemrunes.GemItem;
 import net.fabricmc.fabric.api.client.itemgroup.FabricItemGroupBuilder;
 import net.minecraft.item.ItemGroup;
 import net.minecraft.item.ItemStack;
+import net.minecraft.item.Items;
 import net.minecraft.util.Identifier;
 import org.apache.commons.lang3.tuple.ImmutablePair;
 import org.apache.commons.lang3.tuple.Triple;
@@ -52,5 +53,10 @@ public class CreativeTabs {
         new Identifier(Ref.MODID, "inscribing_group")
         ,
         () -> new ItemStack(ModRegistry.INSCRIBING.ANTI_FIRE));
+
+    public static ItemGroup Tinkering = FabricItemGroupBuilder.build(
+        new Identifier(Ref.MODID, "tinkering_group")
+        ,
+        () -> new ItemStack(Items.CRAFTING_TABLE));
 
 }