+++ src/main/java/com/robertx22/age_of_exile/player_skills/items/alchemy/AlchemyPotionItem.java
@@ -129,7 +129,7 @@ public class AlchemyPotionItem extends Item implements IAutoLocName, IAutoModel,
     }
 
     @Override
-    public PlayerSkillEnum getSkillTypeToCraft() {
+    public PlayerSkillEnum getItemSkillType() {
         return PlayerSkillEnum.ALCHEMY;
     }