+++ src/main/java/com/robertx22/age_of_exile/player_skills/items/foods/SkillItemTier.java
@@ -1,18 +1,20 @@
 package com.robertx22.age_of_exile.player_skills.items.foods;
 
+import com.robertx22.age_of_exile.database.data.level_ranges.LevelRange;
+import com.robertx22.age_of_exile.database.registrators.LevelRanges;
 import net.minecraft.util.Formatting;
 
 import java.util.Arrays;
 
 public enum SkillItemTier {
 
-    SPIRITUAL("Spiritual", 0, 0, 1, Formatting.YELLOW, 60 * 2, 20),
-    CELESTIAL("Celestial", 0.2F, 1, 1.25F, Formatting.AQUA, 60 * 3, 25),
-    EMPYREAN("Empyrean", 0.4F, 2, 1.5F, Formatting.GOLD, 60 * 5, 30),
-    ANGELIC("Angelic", 0.6F, 3, 1.75F, Formatting.LIGHT_PURPLE, 60 * 7, 40),
-    DIVINE("Divine", 0.8F, 4, 2, Formatting.DARK_PURPLE, 60 * 10, 50);
+    SPIRITUAL("Spiritual", LevelRanges.STARTER, 0, 0, 1, Formatting.YELLOW, 60 * 2, 20),
+    CELESTIAL("Celestial", LevelRanges.LOW, 0.2F, 1, 1.25F, Formatting.AQUA, 60 * 3, 25),
+    EMPYREAN("Empyrean", LevelRanges.MIDDLE, 0.4F, 2, 1.5F, Formatting.GOLD, 60 * 5, 30),
+    ANGELIC("Angelic", LevelRanges.HIGH, 0.6F, 3, 1.75F, Formatting.LIGHT_PURPLE, 60 * 7, 40),
+    DIVINE("Divine", LevelRanges.ENDGAME, 0.8F, 4, 2, Formatting.DARK_PURPLE, 60 * 10, 50);
 
-    SkillItemTier(String word, float lvl_req, int tier, float statMulti, Formatting format, int durationseconds, float percent_healed) {
+    SkillItemTier(String word, LevelRange levelRange, float lvl_req, int tier, float statMulti, Formatting format, int durationseconds, float percent_healed) {
         this.word = word;
         this.tier = tier;
         this.statMulti = statMulti;
@@ -20,6 +22,7 @@ public enum SkillItemTier {
         this.percent_healed = percent_healed;
         this.durationSeconds = durationseconds;
         this.lvl_req = lvl_req;
+        this.levelRange = levelRange;
     }
 
     public static SkillItemTier of(int amplifier) {
@@ -29,6 +32,7 @@ public enum SkillItemTier {
             .orElse(SkillItemTier.SPIRITUAL);
     }
 
+    public LevelRange levelRange;
     public float lvl_req;
     public String word;
     public int tier;