+++ src/main/java/com/robertx22/age_of_exile/player_skills/items/mining/MiningStoneItem.java
@@ -1,0 +1,54 @@
+package com.robertx22.age_of_exile.player_skills.items.mining;
+
+import com.robertx22.age_of_exile.aoe_data.datapacks.models.IAutoModel;
+import com.robertx22.age_of_exile.aoe_data.datapacks.models.ItemModelManager;
+import com.robertx22.age_of_exile.database.base.CreativeTabs;
+import com.robertx22.age_of_exile.player_skills.items.foods.SkillItemTier;
+import com.robertx22.age_of_exile.uncommon.interfaces.IAutoLocName;
+import net.minecraft.item.Item;
+import net.minecraft.item.ItemStack;
+import net.minecraft.text.Text;
+import net.minecraft.text.TranslatableText;
+import net.minecraft.util.registry.Registry;
+
+public class MiningStoneItem extends Item implements IAutoLocName, IAutoModel {
+
+    SkillItemTier tier;
+
+    public MiningStoneItem(SkillItemTier tier) {
+        super(new Settings().group(CreativeTabs.Tinkering));
+        this.tier = tier;
+    }
+
+    @Override
+    public Text getName(ItemStack stack) {
+        return new TranslatableText(this.getTranslationKey()).formatted(tier.format);
+    }
+
+    @Override
+    public void generateModel(ItemModelManager manager) {
+        manager.generated(this);
+    }
+
+    @Override
+    public AutoLocGroup locNameGroup() {
+        return AutoLocGroup.Misc;
+    }
+
+    @Override
+    public String locNameLangFileGUID() {
+        return Registry.ITEM.getId(this)
+            .toString();
+    }
+
+    @Override
+    public String locNameForLangFile() {
+        return tier.word + " Stone";
+    }
+
+    @Override
+    public String GUID() {
+        return "stone/" + tier.tier;
+    }
+}
+