+++ src/main/java/com/robertx22/age_of_exile/player_skills/items/tinkering/ChestKeyItem.java
@@ -1,13 +1,11 @@
-package com.robertx22.age_of_exile.player_skills.items.protection_tablets;
+package com.robertx22.age_of_exile.player_skills.items.tinkering;
 
 import com.robertx22.age_of_exile.aoe_data.datapacks.models.IAutoModel;
 import com.robertx22.age_of_exile.aoe_data.datapacks.models.ItemModelManager;
 import com.robertx22.age_of_exile.database.base.CreativeTabs;
 import com.robertx22.age_of_exile.database.data.currency.base.IShapelessRecipe;
 import com.robertx22.age_of_exile.mmorpg.ModRegistry;
-import com.robertx22.age_of_exile.player_skills.IReqSkillLevel;
 import com.robertx22.age_of_exile.player_skills.items.foods.SkillItemTier;
-import com.robertx22.age_of_exile.saveclasses.player_skills.PlayerSkillEnum;
 import com.robertx22.age_of_exile.uncommon.interfaces.IAutoLocName;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
@@ -17,20 +15,29 @@ import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 import net.minecraft.text.LiteralText;
 import net.minecraft.text.Text;
+import net.minecraft.text.TranslatableText;
 import net.minecraft.util.registry.Registry;
 import net.minecraft.world.World;
 
 import java.util.List;
 
-public class ProtectionTabletItem extends Item implements IAutoLocName, IAutoModel, IShapelessRecipe, IReqSkillLevel {
+public class ChestKeyItem extends Item implements IAutoLocName, IAutoModel, IShapelessRecipe {
 
-    TabletTypes type;
-    SkillItemTier tier;
+    public SkillItemTier tier;
 
-    public ProtectionTabletItem(SkillItemTier tier, TabletTypes type) {
-        super(new Settings().group(CreativeTabs.Inscribing));
+    public ChestKeyItem(SkillItemTier tier) {
+        super(new Settings().group(CreativeTabs.Tinkering));
         this.tier = tier;
-        this.type = type;
+    }
+
+    @Override
+    public Text getName(ItemStack stack) {
+        return new TranslatableText(this.getTranslationKey()).formatted(tier.format);
+    }
+
+    @Override
+    public void generateModel(ItemModelManager manager) {
+        manager.generated(this);
     }
 
     @Override
@@ -46,38 +53,19 @@ public class ProtectionTabletItem extends Item implements IAutoLocName, IAutoMod
 
     @Override
     public String locNameForLangFile() {
-        return type.word + " Tablet";
+        return tier.word + " Chest Key";
     }
 
     @Override
     public String GUID() {
-        return "tablet/" + type.id;
-    }
-
-    @Override
-    public void generateModel(ItemModelManager manager) {
-        manager.generated(this);
+        return "key/" + tier.tier;
     }
 
     @Override
     public ShapelessRecipeJsonFactory getRecipe() {
-        ShapelessRecipeJsonFactory fac = ShapelessRecipeJsonFactory.create(this);
-
-        this.type.getRecipeItems()
-            .forEach(x -> fac.input(x));
-        fac.input(ModRegistry.INSCRIBING.INK_TIER_MAP.get(tier));
-
+        ShapelessRecipeJsonFactory fac = ShapelessRecipeJsonFactory.create(this, 3);
+        fac.input(ModRegistry.TINKERING.STONE_TIER_MAP.get(this.tier), 4);
         return fac.criterion("player_level", trigger());
-    }
-
-    @Override
-    public PlayerSkillEnum getItemSkillType() {
-        return PlayerSkillEnum.INSCRIBING;
-    }
-
-    @Override
-    public float getSkillLevelMultiNeeded() {
-        return tier.lvl_req;
     }
 
     @Override
@@ -85,8 +73,9 @@ public class ProtectionTabletItem extends Item implements IAutoLocName, IAutoMod
     public void appendTooltip(ItemStack stack, World world, List<Text> tooltip, TooltipContext context) {
 
         try {
-            tooltip.add(new LiteralText("Keep in Inventory or Ender chest."));
-            tooltip.add(new LiteralText("Activates automatically to save you."));
+
+            tooltip.add(new LiteralText("Opens locked chest items."));
+            tooltip.add(new LiteralText("Needs to be in inventory."));
 
         } catch (Exception e) {
             e.printStackTrace();
@@ -94,3 +83,4 @@ public class ProtectionTabletItem extends Item implements IAutoLocName, IAutoMod
 
     }
 }
+